requirements.txt
Django>=4
djangorestframework==3.13.1
django-filter>=2.4
psycopg2-binary==2.9.3
django-bootstrap4==22.1
fastavro>=1.7
whitenoise==6.0.0
lcogt-logging==0.3.2
hop-client~=0.7
healpy>=1.16
django-cors-headers==3.11.0
mozilla-django-oidc==2.0.0
marshmallow==3.16.0
astropy==5.1
django-extensions # for shell_plus debugging
requests==2.27.1
scramp==1.4.1
jsons==1.6.3 # for Auth serialization
